requirements.txt
Django==3.0.7
django-redis==4.12.1
djangorestframework==3.11.1
gunicorn==20.0.4
pre-commit==2.6.0
psycopg2-binary==2.8.5
pytest==6.0.1
pytest-django==3.9.0
python-redis-lock==3.6.0
redis==3.5.3